8월18일 - 상대방 프로필에서 활동이력엑티비티 들어갔을때, 호스트,게스트 표시 제대로 뜨지않던 것 수정함

+++ app/src/main/java/com/example/abled_food_connect/data/RoomTbDbInfoDataItem.kt
@@ -38,7 +38,7 @@ data class RoomTbDbInfoDataItem(
     @SerializedName("name_host")
     val name_host : String,
     @SerializedName("room_status")
-    val room_status : Int,
+    val room_status : Double,
     @SerializedName("search_keyword")
     val search_keyword : String,
     @SerializedName("map_x")
